Refactor: Move authentication response building to factory class
