+++ vclib-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/OidcSiopWallet.kt
@@ -5,7 +5,6 @@ import at.asitplus.catching
 import at.asitplus.crypto.datatypes.CryptoPublicKey
 import at.asitplus.crypto.datatypes.jws.JsonWebKey
 import at.asitplus.crypto.datatypes.jws.JsonWebKeySet
-import at.asitplus.crypto.datatypes.jws.JweHeader
 import at.asitplus.crypto.datatypes.jws.JwsSigned
 import at.asitplus.crypto.datatypes.jws.toJsonWebKey
 import at.asitplus.crypto.datatypes.pki.leaf
@@ -22,27 +21,25 @@ import at.asitplus.wallet.lib.jws.JwsService
 import at.asitplus.wallet.lib.oidc.OpenIdConstants.Errors
 import at.asitplus.wallet.lib.oidc.OpenIdConstants.ID_TOKEN
 import at.asitplus.wallet.lib.oidc.OpenIdConstants.PREFIX_DID_KEY
-import at.asitplus.wallet.lib.oidc.OpenIdConstants.ResponseMode.*
+import at.asitplus.wallet.lib.oidc.OpenIdConstants.ResponseMode.DIRECT_POST
+import at.asitplus.wallet.lib.oidc.OpenIdConstants.ResponseMode.DIRECT_POST_JWT
 import at.asitplus.wallet.lib.oidc.OpenIdConstants.SCOPE_OPENID
 import at.asitplus.wallet.lib.oidc.OpenIdConstants.URN_TYPE_JWK_THUMBPRINT
 import at.asitplus.wallet.lib.oidc.OpenIdConstants.VP_TOKEN
+import at.asitplus.wallet.lib.oidc.helper.AuthenticationResponseFactory
 import at.asitplus.wallet.lib.oidvci.IssuerMetadata
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception
 import at.asitplus.wallet.lib.oidvci.decodeFromUrlQuery
-import at.asitplus.wallet.lib.oidvci.encodeToParameters
-import at.asitplus.wallet.lib.oidvci.formUrlEncode
 import io.github.aakira.napier.Napier
 import io.ktor.http.*
 import io.ktor.util.*
+import io.ktor.http.Url
+import io.ktor.util.flattenEntries
 import io.matthewnelson.encoding.base16.Base16
-import io.matthewnelson.encoding.base64.Base64
-import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
-import io.matthewnelson.encoding.core.Encoder.Companion.encodeToByteArray
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.datetime.Clock
 import kotlinx.serialization.json.JsonPrimitive
 import kotlinx.serialization.json.buildJsonArray
-import kotlin.random.Random
 import kotlin.time.Duration.Companion.seconds
 
 
@@ -187,105 +184,10 @@ class OidcSiopWallet(
             Napier.w("createAuthnResponse: Unknown response_type ${request.parameters.responseType}")
             throw OAuth2Exception(Errors.INVALID_REQUEST)
         }
-        when (request.parameters.responseMode) {
-            DIRECT_POST -> authnResponseDirectPost(request, response)
-            DIRECT_POST_JWT -> authnResponseDirectPostJwt(request, response)
-            QUERY -> authnResponseQuery(request, response)
-            FRAGMENT, null -> authnResponseFragment(request, response)
-            is OTHER -> TODO()
-        }
-    }
-
-    private fun authnResponseDirectPost(
-        request: AuthenticationRequestParametersFrom<*>,
-        response: AuthenticationResponse
-    ): AuthenticationResponseResult.Post {
-        val url = request.parameters.responseUrl
-            ?: request.parameters.redirectUrl
-            ?: throw OAuth2Exception(Errors.INVALID_REQUEST)
-        return AuthenticationResponseResult.Post(url, response.params.encodeToParameters())
-    }
-
-    /**
-     * Per OID4VP, the response may either be signed, or encrypted (never signed and encrypted!)
-     */
-    private suspend fun authnResponseDirectPostJwt(
-        request: AuthenticationRequestParametersFrom<*>,
-        response: AuthenticationResponse
-    ): AuthenticationResponseResult.Post {
-        val url = request.parameters.responseUrl
-            ?: request.parameters.redirectUrl
-            ?: throw OAuth2Exception(Errors.INVALID_REQUEST)
-        val responseSerialized = buildJarm(request, response)
-        val jarm = AuthenticationResponseParameters(response = responseSerialized)
-        return AuthenticationResponseResult.Post(url, jarm.encodeToParameters())
-    }
-
-    private suspend fun buildJarm(request: AuthenticationRequestParametersFrom<*>, response: AuthenticationResponse) =
-        if (response.clientMetadata != null && response.jsonWebKeys != null && response.clientMetadata.requestsEncryption()) {
-            val alg = response.clientMetadata.authorizationEncryptedResponseAlg!!
-            val enc = response.clientMetadata.authorizationEncryptedResponseEncoding!!
-            val jwk = response.jsonWebKeys.first()
-            val nonce = runCatching { request.parameters.nonce?.decodeToByteArray(Base64()) }.getOrNull()
-                ?: runCatching { request.parameters.nonce?.encodeToByteArray() }.getOrNull()
-                ?: Random.Default.nextBytes(16)
-            val payload = response.params.serialize().encodeToByteArray()
-            jwsService.encryptJweObject(
-                header = JweHeader(
-                    algorithm = alg,
-                    encryption = enc,
-                    type = null,
-                    agreementPartyVInfo = nonce.encodeToByteArray(Base64()),
-                    agreementPartyUInfo = Random.nextBytes(16),
-                    keyId = jwk.keyId,
-                ),
-                payload = payload,
-                recipientKey = jwk,
-                jweAlgorithm = alg,
-                jweEncryption = enc,
-            ).map { it.serialize() }.getOrElse {
-                Napier.w("buildJarm error", it)
-                throw OAuth2Exception(Errors.INVALID_REQUEST)
-            }
-        } else {
-            jwsService.createSignedJwsAddingParams(
-                payload = response.params.serialize().encodeToByteArray(), addX5c = false
-            ).map { it.serialize() }.getOrElse {
-                Napier.w("buildJarm error", it)
-                throw OAuth2Exception(Errors.INVALID_REQUEST)
-            }
-        }
-
-    private fun RelyingPartyMetadata.requestsEncryption() =
-        authorizationEncryptedResponseAlg != null && authorizationEncryptedResponseEncoding != null
-
-    private fun authnResponseQuery(
-        request: AuthenticationRequestParametersFrom<*>,
-        response: AuthenticationResponse
-    ): AuthenticationResponseResult.Redirect {
-        if (request.parameters.redirectUrl == null)
-            throw OAuth2Exception(Errors.INVALID_REQUEST)
-        val url = URLBuilder(request.parameters.redirectUrl).apply {
-            response.params.encodeToParameters().forEach {
-                this.parameters.append(it.key, it.value)
-            }
-        }.buildString()
-        return AuthenticationResponseResult.Redirect(url, response.params)
-    }
-
-    /**
-     * That's the default for `id_token` and `vp_token`
-     */
-    private fun authnResponseFragment(
-        request: AuthenticationRequestParametersFrom<*>,
-        response: AuthenticationResponse
-    ): AuthenticationResponseResult.Redirect {
-        if (request.parameters.redirectUrl == null)
-            throw OAuth2Exception(Errors.INVALID_REQUEST)
-        val url = URLBuilder(request.parameters.redirectUrl)
-            .apply { encodedFragment = response.params.encodeToParameters().formUrlEncode() }
-            .buildString()
-        return AuthenticationResponseResult.Redirect(url, response.params)
+        AuthenticationResponseFactory(jwsService).createAuthenticationResponse(
+            request,
+            response,
+        )
     }
 
     /**
